# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Skadi Public API
  version: 1.0.0
  contact:
    email: zolo@megafoto.sk
  description: The Skadi Public API queries Skadi arbitrage bot and prepares information about the bot to be readily available for public users.
paths:
...
  "/v1/trades":
    get:
      operationId: GetTrades
      summary: Get Arbitrage Trade's informations
      description: Returns an array containing all Arbitrage Trade's informations
      responses:
        "400":
          $ref: '#/components/responses/GeneralErrorResponse'
        "200":
          $ref: '#/components/responses/TradesResponse'
components:
  responses:
...
    TradesResponse:
      description: array containing Arbitrage Trades' informations
      content:
        application/json:
          schema:
            type: array
            $ref: '#/components/schemas/TradeData'
    GeneralErrorResponse:
      description: error message
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    TradeData:
      type: object
      properties:
        time:
          type: string
          description: Trade time
        asset:
          type: string
          description: Asset
        amountIn:
          type: integer
          format: int64
          description: Amount spent in RUNE
        amountOut:
          type: integer
          format: int64
          description: Amount earned in RUNE
        pnl:
          type: string
          description: PNL
    Error:
      required:
        - "error"
      type: object
      properties:
        error:
          type: string
servers:
  - url: http://127.0.0.1:8080
  - url: https://127.0.0.1:8080